schema changes related to employment and education

+++ hfny.jay/Stored Procedures/dbo.spAddEducation.sql
@@ -4,7 +4,7 @@ GO
 SET ANSI_NULLS ON
 GO
 CREATE PROCEDURE [dbo].[spAddEducation](@EducationCreator char(10)=NULL,
-@EducationMonthlyHours smallint=NULL,
+@EducationMonthlyHours int=NULL,
 @FormDate datetime=NULL,
 @FormFK int=NULL,
 @FormType char(2)=NULL,
